requirements.txt
google-genai
streamlit
python-dotenv
streamlit-mic-recorder
Pillow
edge-tts
pygame